Move EntrypointName to opekope2.lilac.annotation
